refactor: config_new.py renamed to config.py
