+++ src/modalities/config/config.py
@@ -1,147 +1,267 @@
-import json
-import warnings
 from enum import Enum
 from pathlib import Path
-from typing import List, Optional, Union
+from typing import Annotated, Any, Dict, List, Optional
 
-from pydantic import BaseModel, Field, FilePath, PositiveFloat, PositiveInt, confloat, conint, model_validator
-from transformers import PretrainedConfig
+import torch.nn as nn
+from pydantic import BaseModel, FilePath, GetCoreSchemaHandler, PositiveInt, conint, validator
+from pydantic_core import core_schema
+from torch.distributed.fsdp import ShardingStrategy
+from torch.optim import Optimizer
+from torch.utils.data import Sampler
+from torch.utils.data.dataset import Dataset
+from transformers import GPT2TokenizerFast
+from transformers.models.llama.tokenization_llama_fast import LlamaTokenizerFast
+from transformers.tokenization_utils_fast import PreTrainedTokenizerFast
 
-from modalities.config.lookup_types import (
-    BatchSamplerTypes,
-    CheckpointingExectionTypes,
-    CheckpointingStrategyTypes,
-    CollatorTypes,
-    DataloaderTypes,
-    DatasetTypes,
-    LossTypes,
-    ModelTypes,
-    OptimizerTypes,
-    SamplerTypes,
-    SchedulerTypes,
-    TokenizerTypes,
-)
-from modalities.config.types import ProcessGroupBackendType
-from modalities.models.gpt2.gpt2_model import GPT2Config
-from modalities.models.huggingface.huggingface_models import HuggingFacePretrainedModelConfig
-from modalities.running_env.fsdp.fsdp_running_env import RunningEnvConfig
-
-
-class WandbConfig(BaseModel):
-    class WandbMode(Enum):
-        ONLINE = "ONLINE"
-        OFFLINE = "OFFLINE"
-        DISABLED = "DISABLED"
-
-    project_name: str
-    mode: WandbMode
-    dir: Optional[Path] = Field(default_factory=lambda: Path("."))
+from modalities.checkpointing.checkpointing import CheckpointingIF
+from modalities.checkpointing.checkpointing_execution import CheckpointingExecutionIF
+from modalities.checkpointing.checkpointing_strategies import CheckpointingStrategyIF
+from modalities.dataloader.dataloader import LLMDataLoader
+from modalities.logging_broker.subscriber import MessageSubscriberIF
+from modalities.loss_functions import Loss
+from modalities.models.gpt2.collator import CollateFnIF
+from modalities.running_env.env_utils import MixedPrecisionSettings, has_bfloat_support
+from modalities.util import parse_enum_by_name
 
 
-class CudaKwargsConfig(BaseModel):
-    num_workers: conint(ge=0)
-    pin_memory: bool
-    shuffle: bool
+class PydanticCheckpointingIF:
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls,
+        _source_type: Any,
+        _handler: GetCoreSchemaHandler,
+    ) -> core_schema.CoreSchema:
+        # see: https://docs.pydantic.dev/latest/concepts/types/#handling-third-party-types
+        return core_schema.json_or_python_schema(
+            json_schema=core_schema.is_instance_schema(CheckpointingIF),
+            python_schema=core_schema.is_instance_schema(CheckpointingIF),
+            # serialization=core_schema.plain_serializer_function_ser_schema(
+            #     lambda instance: instance.x
+            # ),
+        )
 
 
-class TokenizerConfig(BaseModel):
-    class GPT2TokenizerFastConfig(BaseModel):
-        tokenizer_file: str  # FilePath not possible, since transformers.PretrainedTokenizers can only handle strings
-
-    type_hint: TokenizerTypes
-    config: GPT2TokenizerFastConfig
-
-
-class DatasetConfig(BaseModel):
-    class MemMapDatasetConfig(BaseModel):
-        raw_data_path: FilePath
-        index_path: Optional[FilePath] = None
-        block_size: conint(gt=0)
-        tokenizer: TokenizerConfig
-        jq_pattern: str
-        sample_key: str
-
-    class PackedMemMapDatasetContinuousConfig(BaseModel):
-        raw_data_path: Path
-        block_size: conint(gt=0)
-        sample_key: str
-
-    class PackedMemMapDatasetMegatronConfig(BaseModel):
-        raw_data_path: Path
-        block_size: conint(gt=0)
-        sample_key: str
-
-    class MMapIndexedDatasetConfig(BaseModel):
-        path: Path
-        skip_warmup: bool
-
-    class OpenGPTXMMapDatasetConfig(BaseModel):
-        num_samples: conint(ge=1)
-        path: FilePath
-        sample_key: str
-        sequence_len: PositiveInt
-
-    type_hint: DatasetTypes
-    config: Union[
-        MemMapDatasetConfig,
-        OpenGPTXMMapDatasetConfig,
-        PackedMemMapDatasetContinuousConfig,
-        PackedMemMapDatasetMegatronConfig,
-        MMapIndexedDatasetConfig,
-    ] = Field(union_mode="left_to_right")
+class PydanticCheckpointingStrategyIF:
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls,
+        _source_type: Any,
+        _handler: GetCoreSchemaHandler,
+    ) -> core_schema.CoreSchema:
+        # see: https://docs.pydantic.dev/latest/concepts/types/#handling-third-party-types
+        return core_schema.json_or_python_schema(
+            json_schema=core_schema.is_instance_schema(CheckpointingStrategyIF),
+            python_schema=core_schema.is_instance_schema(CheckpointingStrategyIF),
+            # serialization=core_schema.plain_serializer_function_ser_schema(
+            #     lambda instance: instance.x
+            # ),
+        )
 
 
-class SamplerConfig(BaseModel):
-    class DistributedSamplerConfig(BaseModel):
-        rank: conint(ge=0)
-        num_replicas: conint(ge=0)
-        shuffle: bool
-
-    type_hint: SamplerTypes
-    config: DistributedSamplerConfig
-
-
-class BatchSamplerConfig(BaseModel):
-    class StandardBatchSamplerConfig(BaseModel):
-        sampler: SamplerConfig
-        batch_size: conint(gt=0)
-        drop_last: bool
-
-    type_hint: BatchSamplerTypes
-    config: StandardBatchSamplerConfig
+class PydanticCheckpointingExecutionIF:
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls,
+        _source_type: Any,
+        _handler: GetCoreSchemaHandler,
+    ) -> core_schema.CoreSchema:
+        # see: https://docs.pydantic.dev/latest/concepts/types/#handling-third-party-types
+        return core_schema.json_or_python_schema(
+            json_schema=core_schema.is_instance_schema(CheckpointingExecutionIF),
+            python_schema=core_schema.is_instance_schema(CheckpointingExecutionIF),
+            # serialization=core_schema.plain_serializer_function_ser_schema(
+            #     lambda instance: instance.x
+            # ),
+        )
 
 
-class CollatorConfig(BaseModel):
-    class GPT2LLMCollatorConfig(BaseModel):
-        sample_key: str
-        target_key: str
-
-    type_hint: CollatorTypes
-    config: GPT2LLMCollatorConfig
-
-
-class DataLoaderConfig(BaseModel):
-    class LLMDataLoaderConfig(CudaKwargsConfig):
-        dataloader_tag: str
-        dataset: DatasetConfig
-        batch_sampler: BatchSamplerConfig
-        collate_fn: CollatorConfig
-
-    type_hint: DataloaderTypes
-    config: LLMDataLoaderConfig
+class PydanticModelIF:
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls,
+        _source_type: Any,
+        _handler: GetCoreSchemaHandler,
+    ) -> core_schema.CoreSchema:
+        # see: https://docs.pydantic.dev/latest/concepts/types/#handling-third-party-types
+        return core_schema.json_or_python_schema(
+            json_schema=core_schema.is_instance_schema(nn.Module),
+            python_schema=core_schema.is_instance_schema(nn.Module),
+            # serialization=core_schema.plain_serializer_function_ser_schema(
+            #     lambda instance: instance.x
+            # ),
+        )
 
 
-class DataConfig(BaseModel):
-    sample_key: str
-    target_key: str
-    sequence_len: int
-    train_dataloader: DataLoaderConfig
-    eval_dataloaders: List[DataLoaderConfig]
+class PydanticTokenizerIF:
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls,
+        _source_type: Any,
+        _handler: GetCoreSchemaHandler,
+    ) -> core_schema.CoreSchema:
+        # see: https://docs.pydantic.dev/latest/concepts/types/#handling-third-party-types
+        return core_schema.json_or_python_schema(
+            json_schema=core_schema.is_instance_schema(PreTrainedTokenizerFast),
+            python_schema=core_schema.is_instance_schema(PreTrainedTokenizerFast),
+            # serialization=core_schema.plain_serializer_function_ser_schema(
+            #     lambda instance: instance.x
+            # ),
+        )
 
 
-class ModelConfig(BaseModel):
-    type_hint: ModelTypes
-    config: HuggingFacePretrainedModelConfig | GPT2Config
+class PydanticDatasetIF:
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls,
+        _source_type: Any,
+        _handler: GetCoreSchemaHandler,
+    ) -> core_schema.CoreSchema:
+        # see: https://docs.pydantic.dev/latest/concepts/types/#handling-third-party-types
+        return core_schema.json_or_python_schema(
+            json_schema=core_schema.is_instance_schema(Dataset),
+            python_schema=core_schema.is_instance_schema(Dataset),
+            # serialization=core_schema.plain_serializer_function_ser_schema(
+            #     lambda instance: instance.x
+            # ),
+        )
+
+
+class PydanticSamplerIF:
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls,
+        _source_type: Any,
+        _handler: GetCoreSchemaHandler,
+    ) -> core_schema.CoreSchema:
+        # see: https://docs.pydantic.dev/latest/concepts/types/#handling-third-party-types
+        return core_schema.json_or_python_schema(
+            json_schema=core_schema.is_instance_schema(Sampler),
+            python_schema=core_schema.is_instance_schema(Sampler),
+            # serialization=core_schema.plain_serializer_function_ser_schema(
+            #     lambda instance: instance.x
+            # ),
+        )
+
+
+class PydanticCollateFnIF:
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls,
+        _source_type: Any,
+        _handler: GetCoreSchemaHandler,
+    ) -> core_schema.CoreSchema:
+        # see: https://docs.pydantic.dev/latest/concepts/types/#handling-third-party-types
+        return core_schema.json_or_python_schema(
+            json_schema=core_schema.is_instance_schema(CollateFnIF),
+            python_schema=core_schema.is_instance_schema(CollateFnIF),
+            # serialization=core_schema.plain_serializer_function_ser_schema(
+            #     lambda instance: instance.x
+            # ),
+        )
+
+
+class PydanticLLMDataLoaderIF:
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls,
+        _source_type: Any,
+        _handler: GetCoreSchemaHandler,
+    ) -> core_schema.CoreSchema:
+        # see: https://docs.pydantic.dev/latest/concepts/types/#handling-third-party-types
+        return core_schema.json_or_python_schema(
+            json_schema=core_schema.is_instance_schema(LLMDataLoader),
+            python_schema=core_schema.is_instance_schema(LLMDataLoader),
+            # serialization=core_schema.plain_serializer_function_ser_schema(
+            #     lambda instance: instance.x
+            # ),
+        )
+
+
+class PydanticOptimizerIF:
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls,
+        _source_type: Any,
+        _handler: GetCoreSchemaHandler,
+    ) -> core_schema.CoreSchema:
+        # see: https://docs.pydantic.dev/latest/concepts/types/#handling-third-party-types
+        return core_schema.json_or_python_schema(
+            json_schema=core_schema.is_instance_schema(Optimizer),
+            python_schema=core_schema.is_instance_schema(Optimizer),
+            # serialization=core_schema.plain_serializer_function_ser_schema(
+            #     lambda instance: instance.x
+            # ),
+        )
+
+
+class PydanticLossIF:
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls,
+        _source_type: Any,
+        _handler: GetCoreSchemaHandler,
+    ) -> core_schema.CoreSchema:
+        # see: https://docs.pydantic.dev/latest/concepts/types/#handling-third-party-types
+        return core_schema.json_or_python_schema(
+            json_schema=core_schema.is_instance_schema(Loss),
+            python_schema=core_schema.is_instance_schema(Loss),
+            # serialization=core_schema.plain_serializer_function_ser_schema(
+            #     lambda instance: instance.x
+            # ),
+        )
+
+
+class PydanticMessageSubscriberIF:
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls,
+        _source_type: Any,
+        _handler: GetCoreSchemaHandler,
+    ) -> core_schema.CoreSchema:
+        # see: https://docs.pydantic.dev/latest/concepts/types/#handling-third-party-types
+        return core_schema.json_or_python_schema(
+            json_schema=core_schema.is_instance_schema(MessageSubscriberIF),
+            python_schema=core_schema.is_instance_schema(MessageSubscriberIF),
+            # serialization=core_schema.plain_serializer_function_ser_schema(
+            #     lambda instance: instance.x
+            # ),
+        )
+
+
+PydanticCheckpointingIFType = Annotated[CheckpointingIF, PydanticCheckpointingIF]
+PydanticCheckpointingStrategyIFType = Annotated[CheckpointingStrategyIF, PydanticCheckpointingStrategyIF]
+PydanticCheckpointingExecutionIFType = Annotated[CheckpointingExecutionIF, PydanticCheckpointingExecutionIF]
+PydanticModelIFType = Annotated[nn.Module, PydanticModelIF]
+PydanticTokenizerIFType = Annotated[PreTrainedTokenizerFast, PydanticTokenizerIF]
+PydanticDatasetIFType = Annotated[Dataset, PydanticDatasetIF]
+PydanticSamplerIFType = Annotated[Sampler, PydanticSamplerIF]
+PydanticCollateFnIFType = Annotated[CollateFnIF, PydanticCollateFnIF]
+PydanticLLMDataLoaderIFType = Annotated[LLMDataLoader, PydanticLLMDataLoaderIF]
+PydanticOptimizerIFType = Annotated[Optimizer, PydanticOptimizerIF]
+PydanticLossIFType = Annotated[Loss, PydanticLossIF]
+PydanticMessageSubscriberIFType = Annotated[MessageSubscriberIF, PydanticMessageSubscriberIF]
+
+
+class LookupEnum(Enum):
+    @classmethod
+    def _missing_(cls, value: str) -> type:
+        """constructs Enum by member name, if not constructable by value"""
+        return cls.__dict__[value]
+
+
+class TokenizerTypes(LookupEnum):
+    GPT2TokenizerFast = GPT2TokenizerFast
+    LlamaTokenizerFast = LlamaTokenizerFast
+
+
+class PassType(LookupEnum):
+    BY_VALUE = "by_value"
+    BY_REFERENCE = "by_reference"
+
+
+class ReferenceConfig(BaseModel):
+    instance_key: str
+    pass_type: PassType
 
 
 class CLMCrossEntropyLossConfig(BaseModel):
@@ -149,199 +269,227 @@ class CLMCrossEntropyLossConfig(BaseModel):
     prediction_key: str
 
 
-class LossConfig(BaseModel):
-    type_hint: LossTypes
-    config: CLMCrossEntropyLossConfig
+# Checkpointing
 
 
-class TrainingConfig(BaseModel):
-    # TODO: use this in Progress Logging
-    # global_num_training_samples: conint(gt=0)
-    callback_interval_in_samples: conint(gt=0)
-    process_group_backend: ProcessGroupBackendType
-    local_rank: conint(ge=0)
+class SaveEveryKStepsCheckpointingStrategyConfig(BaseModel):
+    k: PositiveInt
+
+
+class SaveKMostRecentCheckpointsStrategyConfig(BaseModel):
+    k: conint(ge=-1)
+
+
+class FSDPToDiscCheckpointingConfig(BaseModel):
+    checkpoint_path: Path
     global_rank: conint(ge=0)
-    world_size: conint(ge=0)
-    main_rank: conint(ge=0)
-    local_train_micro_batch_size: conint(gt=0)
-    global_num_seen_samples: conint(ge=0)
-    do_apply_activation_checkpointing: bool
-    gradient_acc_step: conint(gt=0)
+    experiment_id: str
+    block_names: List[str]
+    mixed_precision_settings: MixedPrecisionSettings
+    sharding_strategy: ShardingStrategy
 
-    @property
-    def local_train_batch_size(self):
-        return self.local_train_micro_batch_size * self.gradient_acc_step
+    @validator("mixed_precision_settings", pre=True, always=True)
+    def parse_mixed_precision_setting_by_name(cls, name):
+        mixed_precision_settings: MixedPrecisionSettings = parse_enum_by_name(
+            name=name, enum_type=MixedPrecisionSettings
+        )
+        if not has_bfloat_support() and (
+            mixed_precision_settings == MixedPrecisionSettings.BF_16
+            or mixed_precision_settings == MixedPrecisionSettings.BF_16_WORKING
+        ):
+            raise ValueError("BF16 not supported in the current environment")
+        return mixed_precision_settings
 
-    @property
-    def global_train_batch_size(self):
-        return self.local_train_batch_size * self.world_size
-
-    # @property
-    # def local_num_train_samples(self):
-    #     exact = self.global_num_training_samples / self.world_size
-    #     ret = self.global_num_training_samples // self.world_size
-    #     if exact != ret:
-    #         print(f"Calculated local_num_training_samples is not an integer. Clipping {exact} to {ret} ")
-    #     return ret
-
-    @property
-    def local_num_seen_train_samples(self):
-        exact = self.global_num_seen_samples / self.world_size
-        ret = self.global_num_seen_samples // self.world_size
-        if exact != ret:
-            print(f"Calculated global_num_seen_samples is not an integer. Clipping {exact} to {ret} ")
-        return ret
-
-    @property
-    def skip_num_local_train_batches(self) -> int:
-        exact = self.global_num_seen_samples / self.world_size / self.local_train_micro_batch_size
-        ret = self.global_num_seen_samples // self.world_size // self.local_train_micro_batch_size
-        if exact != ret:
-            print(f"Calculated skip_num_local_train_batches is not an integer. Clipping {exact} to {ret} ")
-        return ret
-
-    # @property
-    # def num_training_batches(self) -> int:
-    #     exact = self.global_num_training_samples / self.local_train_micro_batch_size
-    #     ret = self.global_num_training_samples // self.local_train_micro_batch_size
-    #     if exact != ret:
-    #         warnings.warn(f"Calculated num_training_batches is not an integer. Clipping {exact} to {ret} ")
-    #     return ret
-
-    @property
-    def callback_interval_in_batches_per_rank(self):
-        exact = self.callback_interval_in_samples / self.local_train_micro_batch_size / self.world_size
-        ret = max(self.callback_interval_in_samples // self.local_train_micro_batch_size // self.world_size, 1)
-        if exact != ret:
-            warnings.warn(
-                f"Calculated callback_interval_in_batches_per_rank is not an integer. Clipping {exact} to {ret} "
-            )
-        return ret
-
-
-class AdamWConfig(BaseModel):
-    lr: confloat(ge=0.0)
-
-
-class OptimizerConfig(BaseModel):
-    type_hint: OptimizerTypes
-    config: AdamWConfig
-
-
-class OneCycleLRConfig(BaseModel):
-    max_lr: PositiveFloat
-    total_steps: conint(ge=1)
-    pct_start: confloat(ge=0.0)
-    anneal_strategy: str
-    cycle_momentum: bool
-    base_momentum: float | List
-    max_momentum: float | List
-    div_factor: PositiveFloat
-    final_div_factor: PositiveFloat
-    three_phase: bool
-    last_epochs: int
-    verbose: bool
-
-
-class StepLRConfig(BaseModel):
-    step_size: conint(ge=1)
-    gamma: confloat(ge=0.0)
-
-
-class ConstantLRConfig(BaseModel):
-    factor: PositiveFloat
-    total_iters: PositiveInt
-
-
-class SchedulerConfig(BaseModel):
-    type_hint: SchedulerTypes
-    config: StepLRConfig | ConstantLRConfig | OneCycleLRConfig
+    @validator("sharding_strategy", pre=True, always=True)
+    def parse_sharding_strategy_by_name(cls, name):
+        return parse_enum_by_name(name=name, enum_type=ShardingStrategy)
 
 
 class CheckpointingConfig(BaseModel):
-    class CheckpointingStrategyConfig(BaseModel):
-        class SaveEveryKStepsCheckpointingStrategyConfig(BaseModel):
-            k: PositiveInt
-
-        class SaveKMostRecentCheckpointsStrategyConfig(BaseModel):
-            k: conint(ge=-1)
-
-        type_hint: CheckpointingStrategyTypes
-        config: SaveEveryKStepsCheckpointingStrategyConfig | SaveKMostRecentCheckpointsStrategyConfig
-
-    class CheckpointingExecutionConfig(BaseModel):
-        class FSDPToDiscCheckpointingConfig(BaseModel):
-            checkpoint_path: Path
-            global_rank: conint(ge=0)
-
-        type_hint: CheckpointingExectionTypes
-        config: FSDPToDiscCheckpointingConfig
-
-    checkpointing_strategy: CheckpointingStrategyConfig
-    checkpointing_execution: CheckpointingExecutionConfig
+    checkpointing_strategy: PydanticCheckpointingStrategyIFType
+    checkpointing_execution: PydanticCheckpointingExecutionIFType
 
 
-class RunMode(Enum):
-    FROM_SCRATCH = "FROM_SCRATCH"
-    WARM_START = "WARM_START"
+class AdamWOptimizerConfig(BaseModel):
+    lr: float
+    wrapped_model: PydanticModelIFType
 
 
-class ModalitiesSetupConfig(BaseModel):
-    class WarmStartSettings(BaseModel):
-        checkpoint_model_path: Path
-        global_num_seen_samples: conint(gt=0)
-        checkpoint_optimizer_path: Optional[Path] = None
-        checkpoint_lr_scheduler_path: Optional[Path] = None
-
-    class FromScratchSettings(BaseModel):
-        global_num_seen_samples: int = 0
-
-    run_mode: RunMode
-    settings: FromScratchSettings
-    # settings: WarmStartSettings
-
-    @model_validator(mode="after")
-    def check_global_num_samples_equal_0_when_from_scratch(self) -> "ModalitiesSetupConfig":
-        if self.run_mode == RunMode.FROM_SCRATCH:
-            if self.settings.global_num_seen_samples != 0:
-                raise ValueError("When starting from scratch, global_num_seen_samples must be 0.")
-        return self
+class CheckpointedOptimizerConfig(BaseModel):
+    checkpointing: PydanticCheckpointingIFType
+    checkpoint_path: Path
+    wrapped_model: PydanticModelIFType
+    optimizer: PydanticOptimizerIFType
 
 
-class AppConfig(BaseModel):
-    modalities_setup: ModalitiesSetupConfig
-    data: DataConfig
-    training: TrainingConfig
-    running_env: RunningEnvConfig
-    model: ModelConfig
-    optimizer: OptimizerConfig
-    scheduler: SchedulerConfig
-    checkpointing: CheckpointingConfig
-    wandb: WandbConfig
-    loss: LossConfig
+class CheckpointedModelConfig(BaseModel):
+    checkpointing: PydanticCheckpointingIFType
+    checkpoint_path: Path
+    model: PydanticModelIFType
 
 
-class PretrainedGPTConfig(PretrainedConfig):
-    model_type = "modalities_gpt2"
+class FSDPWrappedModelConfig(BaseModel):
+    model: PydanticModelIFType
+    sync_module_states: bool
+    mixed_precision_settings: MixedPrecisionSettings
+    sharding_strategy: ShardingStrategy
+    block_names: List[str]
 
-    def __init__(self, config: GPT2Config = None, **kwargs):
-        if type(config) == dict:
-            config = GPT2Config(**config)
-        self.config = config
+    @validator("mixed_precision_settings", pre=True, always=True)
+    def parse_mixed_precision_setting_by_name(cls, name):
+        mixed_precision_settings: MixedPrecisionSettings = parse_enum_by_name(
+            name=name, enum_type=MixedPrecisionSettings
+        )
+        if not has_bfloat_support() and (
+            mixed_precision_settings == MixedPrecisionSettings.BF_16
+            or mixed_precision_settings == MixedPrecisionSettings.BF_16_WORKING
+        ):
+            raise ValueError("BF16 not supported in the current environment")
+        return mixed_precision_settings
 
-        super().__init__(**kwargs)
+    @validator("sharding_strategy", pre=True, always=True)
+    def parse_sharding_strategy_by_name(cls, name):
+        return parse_enum_by_name(name=name, enum_type=ShardingStrategy)
 
-    def to_json_string(self, use_diff: bool = True) -> str:
-        if self.config:
-            json_dict = {"config": self.config.__dict__.copy(), "model_type": self.model_type}
-            json_dict["config"]["attention"] = {
-                "attention_type": self.config.attention.attention_type.value,
-                "scaling_factor": self.config.attention.scaling_factor,
-            }
-            json_dict["config"]["weight_init"] = {
-                "mean": self.config.weight_init.mean,
-                "std": self.config.weight_init.std,
-            }
-        else:
-            json_dict = {}
-        return json.dumps(json_dict)
+
+class GPT2TokenizerFastConfig(BaseModel):
+    # Note: huggingface tokenizers expect file path as string
+    tokenizer_file: str
+
+
+class DistributedSamplerConfig(BaseModel):
+    rank: conint(ge=0)
+    num_replicas: conint(ge=0)
+    shuffle: bool
+    dataset: PydanticDatasetIFType
+
+
+class MemMapDatasetConfig(BaseModel):
+    raw_data_path: FilePath
+    index_path: Optional[FilePath] = None
+    block_size: conint(gt=0)
+    tokenizer: PydanticTokenizerIFType
+    jq_pattern: str
+    sample_key: str
+
+
+class PackedMemMapDatasetContinuousConfig(BaseModel):
+    raw_data_path: Path
+    block_size: conint(gt=0)
+    sample_key: str
+
+
+class PackedMemMapDatasetMegatronConfig(BaseModel):
+    raw_data_path: Path
+    block_size: conint(gt=0)
+    sample_key: str
+
+
+class MMapIndexedDatasetConfig(BaseModel):
+    path: Path
+    skip_warmup: bool
+
+
+class OpenGPTXMMapDatasetConfig(BaseModel):
+    num_samples: conint(ge=1)
+    path: FilePath
+    sample_key: str
+    sequence_len: PositiveInt
+
+
+class BatchSamplerConfig(BaseModel):
+    sampler: PydanticSamplerIF
+    batch_size: conint(gt=0)
+    drop_last: bool
+
+
+class ResumableBatchSamplerConfig(BaseModel):
+    sampler: PydanticSamplerIF
+    start_index: conint(gt=0)
+
+
+class GPT2LLMCollateFnConfig(BaseModel):
+    sample_key: str
+    target_key: str
+
+
+class LLMDataLoaderConfig(BaseModel):
+    dataloader_tag: str
+    dataset: PydanticDatasetIF
+    batch_sampler: PydanticSamplerIF
+    collate_fn: PydanticCollateFnIF
+    num_workers: conint(ge=0)
+    pin_memory: bool
+    shuffle: bool
+    skip_num_batches: Optional[int] = 0
+
+
+class WandbMode(LookupEnum):
+    ONLINE = "ONLINE"
+    OFFLINE = "OFFLINE"
+    DISABLED = "DISABLED"
+
+
+class DummyProgressSubscriberConfig(BaseModel):
+    pass
+
+
+class RichProgressSubscriberConfig(BaseModel):
+    train_dataloader: PydanticLLMDataLoaderIFType
+    eval_dataloaders: Dict[str, PydanticLLMDataLoaderIFType]
+    world_size: int
+    global_num_seen_samples: int
+    local_rank: int
+
+
+class DummyResultSubscriberConfig(BaseModel):
+    pass
+
+
+class WandBEvaluationResultSubscriberConfig(BaseModel):
+    local_rank: int
+    project: str
+    experiment_id: str
+    mode: WandbMode
+    directory: Path
+    experiment_config: Optional[Dict] = None
+
+
+class RichResultSubscriberConfig(BaseModel):
+    num_ranks: int
+    local_rank: int
+
+
+class Settings(BaseModel):
+    class CudaEnv(BaseModel):
+        local_rank: conint(ge=0)
+        world_size: conint(ge=1)
+        global_rank: conint(ge=0)
+
+    class Training(BaseModel):
+        callback_interval_in_batches: conint(ge=1)
+        global_num_training_samples: conint(ge=1)
+        global_num_seen_samples: conint(ge=0)
+        do_apply_activation_checkpointing: bool
+        gradient_acc_step: conint(ge=1)
+        local_train_micro_batch_size: conint(ge=1)
+        sequence_length: conint(ge=1)
+
+    experiment_id: str
+    referencing_keys: Dict[str, str]
+    training: Training
+    cuda_env: CudaEnv
+
+
+class ComponentsModel(BaseModel):
+    wrapped_model: PydanticModelIFType
+    optimizer: PydanticOptimizerIFType
+    loss_fn: PydanticLossIFType
+    train_dataloader: PydanticLLMDataLoaderIFType
+    val_dataloader: PydanticLLMDataLoaderIFType
+    test_dataloader: PydanticLLMDataLoaderIFType
+    batch_progress_subscriber: PydanticMessageSubscriberIFType
+    evaluation_subscriber: PydanticMessageSubscriberIFType
+    checkpointing: PydanticCheckpointingIFType
+
+    settings: Settings
